Add firebase authentication requirement for Favorites and User endpoints

+++ data-at-bat-api/src/main/java/com/dataatbat/data_at_bat_api/services/FavoriteService.java
@@ -1,7 +1,9 @@
 package com.dataatbat.data_at_bat_api.services;
 
 import com.dataatbat.data_at_bat_api.domain.FavoriteEntity;
+import com.dataatbat.data_at_bat_api.domain.TeamEntity;
 import com.dataatbat.data_at_bat_api.persistence.IFavoritesRepository;
+import com.dataatbat.data_at_bat_api.persistence.ITeamsRepository;
 import com.dataatbat.data_at_bat_api.presentation.presentation_models.FavoritesResponse;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
@@ -15,12 +17,14 @@ import java.util.UUID;
 public class FavoriteService {
 
     private final IFavoritesRepository favoritesRepository;
+    public final ITeamsRepository teamsRepository;
 
-    public FavoriteService(IFavoritesRepository favoritesRepository) {
+    public FavoriteService(IFavoritesRepository favoritesRepository, ITeamsRepository teamsRepository) {
         this.favoritesRepository = favoritesRepository;
+        this.teamsRepository = teamsRepository;
     }
 
-    public ResponseEntity<String> createFavorite(UUID teamId, UUID userId) {
+    public ResponseEntity<String> createFavorite(UUID teamId, String userId) {
         try {
             FavoriteEntity favorite = new FavoriteEntity(teamId, userId);
             FavoriteEntity saved = favoritesRepository.save(favorite);
@@ -32,17 +36,59 @@ public class FavoriteService {
         }
     }
 
-    public ResponseEntity<FavoritesResponse> getFavoritesByUserId(UUID userId) {
+    public ResponseEntity<String> createFavorite(String teamAbbreviation, String userId) {
+        try {
+            TeamEntity team = teamsRepository.findByAbbreviation(teamAbbreviation);
+
+            if (team == null) {
+                return ResponseEntity.notFound().build();
+            }
+
+            return createFavorite(team.getTeamId(), userId);
+        } catch (DataIntegrityViolationException e) {
+            return ResponseEntity.badRequest().body("Could not create favorite. Ensure userId and abbreviation are valid.");
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+    public ResponseEntity<FavoritesResponse> getFavoritesByUserId(String userId) {
         List<FavoriteEntity> favorites = favoritesRepository.findByUserId(userId);
         return ResponseEntity.ok(new FavoritesResponse(favorites));
     }
 
-    public ResponseEntity<String> deleteFavorite(UUID id) {
+    public ResponseEntity<String> deleteFavorite(String userId, UUID id) {
         try {
             Optional<FavoriteEntity> favorite = favoritesRepository.findById(id);
-            if (favorite.isEmpty()) return ResponseEntity.notFound().build();
+            if (favorite.isEmpty() || favorite.get().getUserId().equals(userId)) return ResponseEntity.notFound().build();
             favoritesRepository.deleteById(id);
             return ResponseEntity.ok("Favorite deleted successfully.");
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+    public ResponseEntity<String> deleteFavoriteByTeamId(String userId, UUID teamId) {
+        try {
+            FavoriteEntity favorite = favoritesRepository.findByUserIdAndTeamId(userId, teamId);
+            if (favorite == null) {
+                return ResponseEntity.notFound().build();
+            }
+            favoritesRepository.deleteByUserIdAndTeamId(userId, teamId);
+            return ResponseEntity.ok("Favorite deleted successfully.");
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+    public ResponseEntity<String> deleteFavoriteByAbbreviation(String userId, String abbreviation) {
+        try {
+            TeamEntity team = teamsRepository.findByAbbreviation(abbreviation);
+            if (team == null) {
+                return ResponseEntity.notFound().build();
+            }
+
+            return deleteFavoriteByTeamId(userId, team.getTeamId());
         } catch (Exception e) {
             return ResponseEntity.internalServerError().build();
         }
